+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.3 (Not Released Yet)
 
 * Fix composer post update script.
+* Add a config parameter to choose who receives the notifyAccountCreated mail.
 
 ## v0.0.2