+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # CoCOTS Changelog
 
-## v0.0.3 (Not Released Yet)
+## v0.1.0 (Not Released Yet)
 
 * Fix composer post update script.
 * Add a config parameter to choose who receives the notifyAccountCreated mail.
 * New «website title» field.
 * Safer YAML escape for Ansible class.
+* Admin page: column order change.
 
 ## v0.0.2